fix(action): make pesde home dir be in github workspace
The `saveCache` function provided by `@actions/cache` attempts to
resolve paths containing globs and other components relative to the
github workspace. If unable to, it filters the path out. When no paths
are resolved, it throws an error.

In the case of Windows, the `actions/checkout` action writes the
checked out source code to the `D:\` drive. On Linux, it lies in the
home dir. Previously, we used to use `homedir()` to store the pesde
home, which would resolve to a directory within the `C:\` drive.

This makes it impossible for `saveCache` to resolve the pesde home
relative to the workspace dir, hence filtering it out, and having
nothing to cache, only in the special case of Windows.

We now default to `~/.pesde`, and transform components such as the tilde
and references to a POSIX-style HOME variable to be relative to the
github workspace in the `home` input.

FUTURE
======

This issue also led me to discover that it is likely that none of the
package dirs are being cached. Worth investigating.